@@ -1,65 +1,46 @@
 {
-  "name": "pumpfun sniper bot",
+  "name": "meteora-sniper",
   "version": "1.0.0",
-  "description": "Pumpfun sniper that can snipe tokens by dev wallet or ticker...",
-  "main": "src/index.ts",
+  "main": "index.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "ts-node src/index.ts",
     "start": "ts-node src/index.ts",
-    "build": "npx tsc",
-    "encrypt": "npx javascript-obfuscator dist --output dist-encrypt --compact true --control-flow-flattening true --self-defending true"
+    "jsstart": "node dist/index.js",
+    "build": "tsc",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": [
-    "pumpfun",
-    "pumpfun sniper",
-    "trading bot",
-    "pumpdotfun",
-    "pumpdotfun bot"
-  ],
-  "author": "Whistle",
+  "keywords": [],
+  "author": "",
   "license": "ISC",
+  "description": "",
   "dependencies": {
     "@bloxroute/solana-trader-client-ts": "^2.2.2",
     "@coral-xyz/anchor": "^0.30.1",
-    "@cryptoscan/pumpfun-sdk": "^1.3.1",
-    "@metaplex-foundation/js": "^0.20.1",
-    "@metaplex-foundation/mpl-token-metadata": "^3.2.1",
-    "@metaplex-foundation/umi": "^0.9.1",
-    "@raydium-io/raydium-sdk": "^1.3.1-beta.47",
-    "@solana/spl-token": "^0.4.8",
-    "@solana/web3.js": "^1.95.2",
-    "axios": "^1.6.8",
-    "bigint-buffer": "^1.1.5",
+    "@mercurial-finance/dynamic-amm-sdk": "^1.1.23",
+    "@meteora-ag/dlmm": "^1.3.10",
+    "@project-serum/anchor": "^0.26.0",
+    "@raydium-io/raydium-sdk": "^1.3.1-beta.58",
+    "@shyft-to/solana-transaction-parser": "^1.1.20",
+    "@solana/spl-token": "^0.4.9",
+    "@solana/web3.js": "^1.95.5",
+    "@triton-one/yellowstone-grpc": "^1.0.0",
+    "@types/node": "^22.10.0",
+    "axios": "^1.7.8",
     "bn.js": "^5.2.1",
     "bs58": "^6.0.0",
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "dotnet": "^1.1.4",
-    "express": "^4.19.2",
-    "inquirer": "^12.0.1",
-    "jito-ts": "^3.0.1",
-    "json-web-token": "^3.2.0",
+    "inquirer": "^12.1.0",
+    "jito-ts": "^4.1.2",
     "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.5.3",
-    "nodemon": "^3.1.7",
-    "pino": "^8.18.0",
-    "pino-pretty": "^10.3.1",
-    "pino-std-serializers": "^6.2.2",
-    "socket.io": "^4.7.5",
-    "ws": "^8.18.0"
+    "lodash": "^4.17.21",
+    "pino": "^9.6.0",
+    "pino-pretty": "^13.0.0",
+    "tpu-client": "^1.0.6",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.2"
   },
   "devDependencies": {
-    "@types/bn.js": "^5.1.5",
-    "@types/cookie-parser": "^1.4.7",
-    "@types/express": "^4.17.21",
-    "@types/jsonwebtoken": "^9.0.6",
-    "@types/node": "^22.3.0",
-    "javascript-obfuscator": "^4.1.1",
-    "prettier": "^3.2.4",
-    "terser": "^5.37.0",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.5.4"
+    "@types/bn.js": "^5.1.6",
+    "@types/jsonwebtoken": "^9.0.7",
+    "@types/lodash": "^4.17.15"
   }
 }
